change header&hash data type. was: real,double is: log,longint

+++ utils/c/src/shunt_primitives.c
@@ -20,7 +20,7 @@
 //Common Functions 
 ////////////////////////////////////
 
-double shunt_prim_hash(const char *str) {
+long shunt_prim_hash(const char *str) {
   long hash = 5381;
   int c;
